# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Supabase GraphQL API",
    "description": "API for managing companies, establishments, officers, addresses, files, and related data using GraphQL with a Supabase backend. All operations are performed via a single GraphQL endpoint using POST requests.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local GraphQL server"
    }
  ],
  "paths": {
...
    "/api/pronto/personas": {
      "get": { 
        "summary": "List all personas", 
        "description": "Retrieves a list of all personas available in the system",
        "tags": ["Pronto"], 
        "responses": { 
          "200": { 
            "description": "List of personas", 
            "content": { 
              "application/json": { 
                "schema": { 
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "personas": {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/ProntoPersona"
                        }
                      }
                    }
                  }
                },
                "example": [
                  {
                    "personas": [
                      {
                        "uuid": "8e1dcba0-f0bb-4071-99fc-f18ba6559ccc",
                        "name": "VP Grenoble",
                        "job_titles": ["VP of Engineering", "CTO"],
                        "location": ["Grenoble, Auvergne-Rhône-Alpes, France"],
                        "created_at": "2024-09-09T15:15:15.951Z",
                        "updated_at": "2024-09-09T15:15:15.951Z"
                      }
                    ]
                  }
                ]
              } 
            } 
          },
          "401": {
            "description": "Unauthorized - Invalid API key"
          },
          "500": {
            "description": "Internal server error"
          }
        } 
      }
    },
    "/api/pronto/personas/{id}": {
      "get": { 
        "summary": "Get a specific persona", 
        "description": "Retrieves details of a specific persona by UUID",
        "tags": ["Pronto"], 
        "parameters": [ 
          { 
            "name": "id", 
            "in": "path", 
            "required": true, 
            "schema": { "type": "string", "format": "uuid" },
            "description": "UUID of the persona to retrieve"
          } 
        ], 
        "responses": { 
          "200": { 
            "description": "Persona details", 
            "content": { 
              "application/json": { 
                "schema": { "$ref": "#/components/schemas/ProntoPersona" },
                "example": {
                  "uuid": "8e1dcba0-f0bb-4071-99fc-f18ba6559ccc",
                  "name": "VP Grenoble",
                  "job_titles": ["VP of Engineering", "CTO"],
                  "location": ["Grenoble, Auvergne-Rhône-Alpes, France"],
                  "created_at": "2024-09-09T15:15:15.951Z",
                  "updated_at": "2024-09-09T15:15:15.951Z"
                }
              } 
            } 
          }, 
          "404": { "description": "Persona not found" },
          "401": { "description": "Unauthorized - Invalid API key" },
          "500": { "description": "Internal server error" }
        } 
      }
    },
    "/api/pronto/lists": {
      "get": { 
        "summary": "Retrieve all lists", 
        "description": "Retrieves a list of all contact and company lists available in the system",
        "tags": ["Pronto"], 
        "responses": { 
          "200": { 
            "description": "List of lists", 
            "content": { 
              "application/json": { 
                "schema": { 
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProntoList"
                  }
                },
                "example": [
                  {
                    "uuid": "123e4567-e89b-12d3-a456-426614174000",
                    "name": "Tech Companies",
                    "description": "List of technology companies",
                    "created_at": "2024-01-01T10:00:00.000Z",
                    "updated_at": "2024-01-01T10:00:00.000Z"
                  }
                ]
              } 
            } 
          },
          "401": { "description": "Unauthorized - Invalid API key" },
          "500": { "description": "Internal server error" }
        } 
      },
      "post": { 
        "summary": "Create a new list", 
        "description": "Creates a new list for organizing contacts and companies",
        "tags": ["Pronto"], 
        "requestBody": { 
          "required": true, 
          "content": { 
            "application/json": { 
              "schema": { 
                "type": "object",
                "properties": {
                  "name": { "type": "string", "description": "Name of the list" },
                  "description": { "type": "string", "description": "Description of the list" }
                },
                "required": ["name"]
              },
              "example": {
                "name": "New Tech List",
                "description": "List of new technology companies"
              }
            } 
          } 
        }, 
        "responses": { 
          "201": { 
            "description": "List created", 
            "content": { 
              "application/json": { 
                "schema": { "$ref": "#/components/schemas/ProntoList" },
                "example": {
                  "uuid": "123e4567-e89b-12d3-a456-426614174001",
                  "name": "New Tech List",
                  "description": "List of new technology companies",
                  "created_at": "2024-01-01T10:00:00.000Z",
                  "updated_at": "2024-01-01T10:00:00.000Z"
                }
              } 
            } 
          }, 
          "400": { "description": "Invalid input" },
          "401": { "description": "Unauthorized - Invalid API key" },
          "500": { "description": "Internal server error" }
        } 
      }
    },
    "/api/pronto/lists/{id}": {
      "get": { 
        "summary": "Retrieve a specific list", 
        "description": "Retrieves details of a specific list by UUID",
        "tags": ["Pronto"], 
        "parameters": [ 
          { 
            "name": "id", 
            "in": "path", 
            "required": true, 
            "schema": { "type": "string", "format": "uuid" },
            "description": "UUID of the list to retrieve"
          } 
        ], 
        "responses": { 
          "200": { 
            "description": "List details", 
            "content": { 
              "application/json": { 
                "schema": { "$ref": "#/components/schemas/ProntoList" },
                "example": {
                  "uuid": "123e4567-e89b-12d3-a456-426614174000",
                  "name": "Tech Companies",
                  "description": "List of technology companies",
                  "created_at": "2024-01-01T10:00:00.000Z",
                  "updated_at": "2024-01-01T10:00:00.000Z"
                }
              } 
            } 
          }, 
          "404": { "description": "List not found" },
          "401": { "description": "Unauthorized - Invalid API key" },
          "500": { "description": "Internal server error" }
        } 
      },
      "put": { 
        "summary": "Update a list", 
        "description": "Updates an existing list with new information",
        "tags": ["Pronto"], 
        "parameters": [ 
          { 
            "name": "id", 
            "in": "path", 
            "required": true, 
            "schema": { "type": "string", "format": "uuid" },
            "description": "UUID of the list to update"
          } 
        ], 
        "requestBody": { 
          "required": true, 
          "content": { 
            "application/json": { 
              "schema": { 
                "type": "object",
                "properties": {
                  "name": { "type": "string", "description": "Name of the list" },
                  "description": { "type": "string", "description": "Description of the list" }
                }
              },
              "example": {
                "name": "Updated Tech List",
                "description": "Updated list of technology companies"
              }
            } 
          } 
        }, 
        "responses": { 
          "200": { 
            "description": "List updated", 
            "content": { 
              "application/json": { 
                "schema": { "$ref": "#/components/schemas/ProntoList" },
                "example": {
                  "uuid": "123e4567-e89b-12d3-a456-426614174000",
                  "name": "Updated Tech List",
                  "description": "Updated list of technology companies",
                  "created_at": "2024-01-01T10:00:00.000Z",
                  "updated_at": "2024-01-02T10:00:00.000Z"
                }
              } 
            } 
          }, 
          "400": { "description": "Invalid input" }, 
          "404": { "description": "List not found" },
          "401": { "description": "Unauthorized - Invalid API key" },
          "500": { "description": "Internal server error" }
        } 
      }
    },
    "/api/pronto/searches": {
      "get": { 
        "summary": "Retrieve all searches", 
        "description": "Retrieves a list of all saved searches in the system",
        "tags": ["Pronto"], 
        "responses": { 
          "200": { 
            "description": "List of searches", 
            "content": { 
              "application/json": { 
                "schema": { 
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProntoSearch"
                  }
                },
                "example": [
                  {
                    "uuid": "123e4567-e89b-12d3-a456-426614174000",
                    "name": "Tech Companies in Paris",
                    "query": {
                      "industry": "Technology",
                      "location": "Paris, France",
                      "size": "10-50"
                    },
                    "created_at": "2024-01-01T10:00:00.000Z",
                    "updated_at": "2024-01-01T10:00:00.000Z"
                  }
                ]
              } 
            } 
          },
          "401": { "description": "Unauthorized - Invalid API key" },
          "500": { "description": "Internal server error" }
        } 
      }
    },
    "/api/pronto/searches/{id}": {
      "get": { 
        "summary": "Retrieve search details", 
        "description": "Retrieves details of a specific saved search by UUID",
        "tags": ["Pronto"], 
        "parameters": [ 
          { 
            "name": "id", 
            "in": "path", 
            "required": true, 
            "schema": { "type": "string", "format": "uuid" },
            "description": "UUID of the search to retrieve"
          } 
        ], 
        "responses": { 
          "200": { 
            "description": "Search details", 
            "content": { 
              "application/json": { 
                "schema": { "$ref": "#/components/schemas/ProntoSearch" },
                "example": {
                  "uuid": "123e4567-e89b-12d3-a456-426614174000",
                  "name": "Tech Companies in Paris",
                  "query": {
                    "industry": "Technology",
                    "location": "Paris, France",
                    "size": "10-50"
                  },
                  "created_at": "2024-01-01T10:00:00.000Z",
                  "updated_at": "2024-01-01T10:00:00.000Z"
                }
              } 
            } 
          }, 
          "404": { "description": "Search not found" },
          "401": { "description": "Unauthorized - Invalid API key" },
          "500": { "description": "Internal server error" }
        } 
      }
    },
    "/api/pronto/searches/{id}/leads": {
      "get": { 
        "summary": "Retrieve leads from a specific search", 
        "description": "Retrieves all leads/contacts from a specific saved search with pagination support",
        "tags": ["Pronto"], 
        "parameters": [ 
          { 
            "name": "id", 
            "in": "path", 
            "required": true, 
            "schema": { "type": "string", "format": "uuid" },
            "description": "UUID of the search to retrieve leads from"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 1, "minimum": 1 },
            "description": "Page number for pagination"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 100, "minimum": 1, "maximum": 1000 },
            "description": "Number of leads per page"
          }
        ], 
        "responses": { 
          "200": { 
            "description": "Leads from the search", 
            "content": { 
              "application/json": { 
                "schema": { 
                  "type": "object",
                  "properties": {
                    "leads": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProntoContact"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": { "type": "integer", "description": "Current page number" },
                        "limit": { "type": "integer", "description": "Number of items per page" },
                        "total": { "type": "integer", "description": "Total number of leads" },
                        "pages": { "type": "integer", "description": "Total number of pages" }
                      }
                    }
                  }
                },
                "example": {
                  "leads": [
                    {
                      "uuid": "123e4567-e89b-12d3-a456-426614174000",
                      "first_name": "John",
                      "last_name": "Doe",
                      "email": "john.doe@company.com",
                      "job_title": "CEO",
                      "company": "Tech Corp",
                      "linkedin_url": "https://linkedin.com/in/johndoe",
                      "phone": "+33123456789"
                    }
                  ],
                  "pagination": {
                    "page": 1,
                    "limit": 100,
                    "total": 60,
                    "pages": 1
                  }
                }
              } 
            } 
          }, 
          "404": { "description": "Search not found" },
          "401": { "description": "Unauthorized - Invalid API key" },
          "500": { "description": "Internal server error" }
        } 
      }
    },
    "/api/pronto/accounts/profiles": {
      "post": { 
        "summary": "Get profiles in a company", 
        "description": "Retrieves employee profiles and contact information for a specific company",
        "tags": ["Pronto"], 
        "requestBody": { 
          "required": true, 
          "content": { 
            "application/json": { 
              "schema": { 
                "type": "object",
                "properties": {
                  "company_name": { "type": "string", "description": "Name of the company" },
                  "domain": { "type": "string", "description": "Company domain/website" },
                  "limit": { "type": "integer", "description": "Maximum number of profiles to return" }
                },
                "required": ["company_name"]
              },
              "example": {
                "company_name": "Google",
                "domain": "google.com",
                "limit": 10
              }
            } 
          } 
        }, 
        "responses": { 
          "200": { 
            "description": "Company profiles", 
            "content": { 
              "application/json": { 
                "schema": { 
                  "type": "object",
                  "properties": {
                    "profiles": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProntoContact"
                      }
                    },
                    "total": { "type": "integer", "description": "Total number of profiles found" }
                  }
                },
                "example": {
                  "profiles": [
                    {
                      "uuid": "123e4567-e89b-12d3-a456-426614174000",
                      "first_name": "John",
                      "last_name": "Doe",
                      "email": "john.doe@google.com",
                      "job_title": "Software Engineer",
                      "company": "Google",
                      "linkedin_url": "https://linkedin.com/in/johndoe"
                    }
                  ],
                  "total": 1
                }
              } 
            } 
          }, 
          "400": { "description": "Invalid input" },
          "401": { "description": "Unauthorized - Invalid API key" },
          "500": { "description": "Internal server error" }
        } 
      }
    },
    "/api/pronto/accounts/headcount": {
      "post": { 
        "summary": "Get headcount details", 
        "description": "Retrieves detailed headcount information and employee count evolution for a company",
        "tags": ["Pronto"], 
        "requestBody": { 
          "required": true, 
          "content": { 
            "application/json": { 
              "schema": { 
                "type": "object",
                "properties": {
                  "company_name": { "type": "string", "description": "Name of the company" },
                  "domain": { "type": "string", "description": "Company domain/website" }
                },
                "required": ["company_name"]
              },
              "example": {
                "company_name": "Google",
                "domain": "google.com"
              }
            } 
          } 
        }, 
        "responses": { 
          "200": { 
            "description": "Headcount details", 
            "content": { 
              "application/json": { 
                "schema": { 
                  "type": "object",
                  "properties": {
                    "company": { "type": "string", "description": "Company name" },
                    "current_headcount": { "type": "integer", "description": "Current number of employees" },
                    "headcount_history": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "date": { "type": "string", "format": "date", "description": "Date of headcount record" },
                          "count": { "type": "integer", "description": "Number of employees on that date" }
                        }
                      }
                    }
                  }
                },
                "example": {
                  "company": "Google",
                  "current_headcount": 150000,
                  "headcount_history": [
                    {
                      "date": "2023-01-01",
                      "count": 140000
                    },
                    {
                      "date": "2024-01-01",
                      "count": 150000
                    }
                  ]
                }
              } 
            } 
          }, 
          "400": { "description": "Invalid input" },
          "401": { "description": "Unauthorized - Invalid API key" },
          "500": { "description": "Internal server error" }
        } 
      }
    },
    "/api/pronto/accounts/extract": {
      "post": { 
        "summary": "Extract account search results", 
        "description": "Extracts company data based on search criteria and filters",
        "tags": ["Pronto"], 
        "requestBody": { 
          "required": true, 
          "content": { 
            "application/json": { 
              "schema": { 
                "type": "object",
                "properties": {
                  "query": { "type": "string", "description": "Search query for companies" },
                  "filters": {
                    "type": "object",
                    "properties": {
                      "industry": { "type": "string", "description": "Industry filter" },
                      "location": { "type": "string", "description": "Location filter" },
                      "size": { "type": "string", "description": "Company size filter" },
                      "funding": { "type": "string", "description": "Funding status filter" }
                    }
                  },
                  "limit": { "type": "integer", "description": "Maximum number of results to return" }
                },
                "required": ["query"]
              },
              "example": {
                "query": "technology companies",
                "filters": {
                  "industry": "Technology",
                  "location": "Paris, France",
                  "size": "10-50"
                },
                "limit": 10
              }
            } 
          } 
        }, 
        "responses": { 
          "200": { 
            "description": "Account search results", 
            "content": { 
              "application/json": { 
                "schema": { 
                  "type": "object",
                  "properties": {
                    "accounts": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProntoAccount"
                      }
                    },
                    "total": { "type": "integer", "description": "Total number of accounts found" }
                  }
                },
                "example": {
                  "accounts": [
                    {
                      "uuid": "123e4567-e89b-12d3-a456-426614174000",
                      "name": "TechCorp",
                      "website": "https://techcorp.com",
                      "linkedin_url": "https://linkedin.com/company/techcorp",
                      "industry": "Technology",
                      "size": "10-50",
                      "location": "Paris, France"
                    }
                  ],
                  "total": 1
                }
              } 
            } 
          }, 
          "400": { "description": "Invalid input" },
          "401": { "description": "Unauthorized - Invalid API key" },
          "500": { "description": "Internal server error" }
        } 
      }
    },
...
    "/api/pronto/enrichments/account": {
      "post": { 
        "summary": "Single account enrichment", 
        "description": "Enriches a single company account with additional firmographic data",
        "tags": ["Pronto"], 
        "requestBody": { 
          "required": true, 
          "content": { 
            "application/json": { 
              "schema": { 
                "type": "object",
                "properties": {
                  "company_name": { "type": "string", "description": "Name of the company" },
                  "domain": { "type": "string", "description": "Company domain/website" },
                  "linkedin_url": { "type": "string", "description": "Company LinkedIn URL" }
                },
                "required": ["company_name"]
              },
              "example": {
                "company_name": "TechCorp",
                "domain": "techcorp.com",
                "linkedin_url": "https://linkedin.com/company/techcorp"
              }
            } 
          } 
        }, 
        "responses": { 
          "200": { 
            "description": "Enriched account data", 
            "content": { 
              "application/json": { 
                "schema": { "$ref": "#/components/schemas/ProntoAccount" },
                "example": {
                  "uuid": "123e4567-e89b-12d3-a456-426614174000",
                  "name": "TechCorp",
                  "website": "https://techcorp.com",
                  "linkedin_url": "https://linkedin.com/company/techcorp",
                  "industry": "Technology",
                  "size": "10-50",
                  "location": "Paris, France"
                }
              } 
            } 
          }, 
          "400": { "description": "Invalid input" },
          "401": { "description": "Unauthorized - Invalid API key" },
          "500": { "description": "Internal server error" }
        } 
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProntoPersona": {
        "type": "object",
        "properties": {
          "uuid": { "type": "string", "format": "uuid", "description": "Unique identifier for the persona" },
          "name": { "type": "string", "description": "Name of the persona" },
          "job_titles": { 
            "type": "array", 
            "items": { "type": "string" }, 
            "description": "List of job titles for this persona" 
          },
          "location": { 
            "type": "array", 
            "items": { "type": "string" }, 
            "description": "List of locations for this persona" 
          },
          "created_at": { "type": "string", "format": "date-time", "description": "Creation timestamp" },
          "updated_at": { "type": "string", "format": "date-time", "description": "Last update timestamp" }
        },
        "required": ["uuid", "name"]
      },
      "ProntoList": {
        "type": "object",
        "properties": {
          "uuid": { "type": "string", "format": "uuid", "description": "Unique identifier for the list" },
          "name": { "type": "string", "description": "Name of the list" },
          "description": { "type": "string", "description": "Description of the list" },
          "created_at": { "type": "string", "format": "date-time", "description": "Creation timestamp" },
          "updated_at": { "type": "string", "format": "date-time", "description": "Last update timestamp" }
        },
        "required": ["uuid", "name"]
      },
...
    }
  },
...
}